Move wookie Token spec into top-level wookie package
