+++ pkg/authz/wookie/model.go
@@ -14,13 +14,17 @@
 
 package authz
 
-import "time"
+import (
+	"time"
+
+	"github.com/warrant-dev/warrant/pkg/wookie"
+)
 
 type Model interface {
 	GetID() int64
 	GetVersion() int64
 	GetCreatedAt() time.Time
-	ToToken() *Token
+	ToToken() *wookie.Token
 }
 
 type Wookie struct {
@@ -41,8 +45,8 @@ func (w Wookie) GetCreatedAt() time.Time {
 	return w.CreatedAt
 }
 
-func (w Wookie) ToToken() *Token {
-	return &Token{
+func (w Wookie) ToToken() *wookie.Token {
+	return &wookie.Token{
 		ID:        w.ID,
 		Version:   w.Version,
 		Timestamp: w.CreatedAt,
