+++ pkg/wookie/token.go
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package authz
+package wookie
 
 import (
 	"encoding/base64"
